borer fixes

+++ code/modules/mob/living/simple_animal/borer/borer_egg.dm
@@ -7,19 +7,23 @@
 	icon_state = "borer_egg"
 	var/health = 100
 	var/hitnoise = 'sound/items/Welder.ogg'
+	var/timeleft
 	layer = TURF_LAYER
 
 /obj/structure/borer_egg/New()
 	..()
-	spawn(rand(1800, 3000))
-		processing_objects.Add(src)
+	timeleft = world.time + rand(1800, 3000)
+	processing_objects.Add(src)
 
 /obj/structure/borer_egg/process()
+
+	if(timeleft > world.time)
+		return
 
 	var/mob/dead/observer/ghost
 	for(var/mob/dead/observer/O in src.loc)
 		if(!O.client)	continue
-		if(O.mind && O.mind.current && O.mind.current.stat != DEAD)	continue
+		if(O.can_reenter_corpse && O.mind && O.mind.current && O.mind.current.stat != DEAD)	continue
 		ghost = O
 		break
 
@@ -27,10 +31,14 @@
 	var/datum/gas_mixture/environment = location.return_air()
 	if(environment.toxins >= MOLES_PLASMA_VISIBLE && ghost)
 		var/response = alert(ghost, "This egg is ready to hatch. Do you want to be a cortical borer?",, "Yes", "No")
+		if(!src || src.gc_destroyed)
+			return
 		if(response == "Yes")
 			var/mob/living/simple_animal/borer/B = new /mob/living/simple_animal/borer
 			B.loc = src.loc
 			B.key = ghost.key
+			if(B.evil)
+				B.make_special()
 			qdel(src)
 	else if(environment.toxins >= MOLES_PLASMA_VISIBLE || ghost)
 		if(icon_state != "borer_egg_pulsing")
